Implement Ns_DStringVPrintf() using OS vsnprintf() to pick up modern format specifiers and bug fixes
(dsprintf.c was 7 year old NetBSD code)

This is similar to the old AOLserver 3.x implementation except that
instead of printing into a large stack buffer and ignoring overflows,
taking the hit of always copying the result into the dstring, we
manipulate the dstring buffer and print directly into it. (There are
only 2 instances in the current test suite where we overflow static
space and need to grow.)

NB: Win32 is handled, but untested. I don't think we care about Irix any
more, which seems to have been a concern of the old code.

* nsd/tcltime.c: Quieten some compiler warnings when platform time_t is
not a long.

+++ nsd/sched.c
@@ -888,9 +888,10 @@ NsGetScheduled(Tcl_DString *dsPtr)
     while (hPtr != NULL) {
         ePtr = Tcl_GetHashValue(hPtr);
         Tcl_DStringStartSublist(dsPtr);
-        Ns_DStringPrintf(dsPtr, "%d %d %d %ld %ld %ld %ld",
-            ePtr->id, ePtr->flags, ePtr->interval, ePtr->nextqueue,
-            ePtr->lastqueue, ePtr->laststart, ePtr->lastend);
+        Ns_DStringPrintf(dsPtr, "%d %d %d %jd %jd %jd %jd",
+            ePtr->id, ePtr->flags, ePtr->interval,
+            (intmax_t) ePtr->nextqueue, (intmax_t) ePtr->lastqueue,
+            (intmax_t) ePtr->laststart, (intmax_t) ePtr->lastend);
         Ns_GetProcInfo(dsPtr, ePtr->proc, ePtr->arg);
         Tcl_DStringEndSublist(dsPtr);
         hPtr = Tcl_NextHashEntry(&search);
